Lab_2/Ejercicio_Final/codigo_ejercicioFinal/requirements.txt
pyserial==3.5
numpy==1.24.3
opencv-python==4.8.0.76
requests>=2.25.0
Pillow>=8.0.0 